pyproject.toml
[project]
name = "retake"
version = "1.0.0"
description = "Video retake editor with voice cloning"
requires-python = ">=3.14"
dependencies = [
    "moviepy>=2.1.1",
    "opencv-python>=4.10.0",
    "ffmpeg-python>=0.2.0",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "numpy>=2.1.3",
    "scipy>=1.14.1",
    "soundfile>=0.12.1",
    "pillow>=10.4.0",
    "tqdm>=4.66.6",
    "pyyaml>=6.0.2",
    "imageio>=2.35.1",
    "imageio-ffmpeg>=0.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cpu"]
